rag_app/backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
openai==1.12.0
chromadb==0.4.22
PyMuPDF==1.23.26
python-docx==1.1.0
markdown==3.5.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
tavily-python==0.3.1